Refactor `ExportQueryPlugin` to inherit from new `SplitgraphExportPlugin` base class
Adopt the same pattern used for import plugin inheritance, so that
we can create a plugin for exporting to Seafowl (as opposed to exporting
to a parquet/csv file, like the current plugin), while re-using most of
the code, particularly regarding waiting for jobs.
